requirements.txt
fastapi==0.111.0
uvicorn==0.30.0
pydantic==2.8.0
python-dotenv==1.0.1

langgraph==0.2.56
langchain-core==0.3.49
langchain-groq==0.2.5
chromadb>=1.5.4,<2.0.0
sentence-transformers==5.2.3
numpy==2.1.1

tavily-python==0.3.2

streamlit==1.50.0
requests==2.32.3

matplotlib==3.10.1

langsmith==0.1.147
opentelemetry-api==1.25.0
opentelemetry-sdk==1.25.0
opentelemetry-exporter-otlp==1.25.0
